requirements.txt
# Core Flask and API
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# OCR and Image Processing
pytesseract==0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
pdf2image>=1.16.0
numpy>=1.24.0
scikit-image>=0.21.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.9.0

# Text Processing and NLP
regex>=2023.8.0
python-dateutil>=2.8.0
fuzzywuzzy>=0.18.0

# Utilities
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0

# Development
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
